+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Changed
+- `JSONAuto`: added `JSONConfig` to functions
+- `JSONDeserializer`: added `JSONConfig` to functions
 
 ## [0.5] - 2019-05-21
 ### Changed